+++ BHoM_Adapter/IO/Extra/Execute.cs
@@ -1,0 +1,24 @@
+﻿using BH.Engine.Reflection;
+using BH.oM.Base;
+using BH.oM.Data.Collections;
+using BH.oM.Data.Requests;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace BH.Adapter
+{
+    public abstract partial class BHoMAdapter
+    {
+
+        /***************************************************/
+        /**** Public Adapter Methods                    ****/
+        /***************************************************/
+
+        public virtual bool Execute(string command, Dictionary<string, object> parameters = null, Dictionary<string, object> config = null)
+        {
+            return false;
+        }
+    }
+}
